fix: made scheduler_task and isolated to make adding other options easier
